feat(o-agent): fix str_replace_editor view_range bug
